requirements.txt
matplotlib==3.9.2
numpy==2.1.3
python_docx==1.1.2
PyYAML==6.0.2
